requirements.txt
flask 
flask-jsonpify
flask-sqlalchemy
flask-restful
asterisk-ami
